[Model] Update and optimize the Qwen2.5-VL model (#572)
Update and optimize the Qwen2.5-VL model.
1. Support Qwen2.5-VL-32B and add the sample configuration.
2. Optimize the data processing process to decrease the time consumed.
3. Other optimizations, such as memory cache, recompute strategy,
distributed parallelism, and so on.

🚀 **QuickStart Available**:  
Start your SFT training quickly by referring to the [QuickStart
Guide](https://github.com/FlagOpen/FlagScale/pull/572/files#diff-1dc2d875e1726124f3b3ba085e5ed5332d68cb088e8d3473d59d55c35bd725f1).

Based on FlagScale, BAAI has achieved
[RoboBrain2.0](https://superrobobrain.github.io/), the most powerful
open-source embodied brain model to date.
PS: Adapted from
https://github.com/alibaba/Pai-Megatron-Patch/tree/4c305eff70c5c1d30d65ceed2713acabff48ef87/examples/qwen2_5_vl
and
https://github.com/huggingface/transformers/tree/10627c1a0f6877ce6715b9537afe7fafb2a89edd/src/transformers/models/qwen2_5_vl

---------

+++ flagscale/train/models/qwen2_5_vl/vit_model.py
@@ -62,14 +62,15 @@ class VisionRotaryEmbedding(nn.Module):
     """
     def __init__(self, dim: int, theta: float = 10000.0) -> None:
         super().__init__()
-        inv_freq = 1.0 / (theta ** (torch.arange(0, dim, 2, dtype=torch.float) / dim))
+        # NOTE(lizhiyu): print inv_freq to check it.
+        inv_freq = 1.0 / (theta ** (torch.arange(0, dim, 2, dtype=torch.bfloat16) / dim))
         self.register_buffer("inv_freq", inv_freq, persistent=False)
 
     def forward(self, seqlen: int) -> torch.Tensor:
         seq = torch.arange(seqlen, device=self.inv_freq.device, dtype=self.inv_freq.dtype)
         # freqs [seq_len, dim // 2]
         freqs = torch.outer(seq, self.inv_freq)
-        return freqs.float()
+        return freqs
 
 # reference from https://github.com/huggingface/transformers/blob/0ad3710d4767d4ac7ee95f33f8554373e59efade/src/transformers/models/qwen2_5_vl/modular_qwen2_5_vl.py#L243
 class Qwen2_5VisionModel(VisionModule):
